[Kinetics] Move TwoTempPlasma to its own header file
